+++ Assets/RefactorExamplesData/ComponentD.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace RefactorExamplesData
+{
+    public class ComponentD : MonoBehaviour
+    {
+        public int someValue;
+    }
+}